+++ kedro_snowflake/starters/snowflights/{{ cookiecutter.repo_name }}/src/{{ cookiecutter.python_package }}/settings.py
@@ -1,10 +1,11 @@
 """Project settings. There is no need to edit this file unless you want to change values
 from the Kedro defaults. For further information, including these default values, see
 https://kedro.readthedocs.io/en/stable/kedro_project_setup/settings.html."""
+from .hooks import SnowflakeHook
 
 # Instantiated project hooks.
-# from {{cookiecutter.python_package}}.hooks import ProjectHooks
-# HOOKS = (ProjectHooks(),)
+# from snowflake_mlops_demo.hooks import ProjectHooks
+HOOKS = (SnowflakeHook(),)
 
 # Installed plugins for which to disable hook auto-registration.
 # DISABLE_HOOKS_FOR_PLUGINS = ("kedro-viz",)
@@ -30,9 +31,7 @@ from kedro.config import OmegaConfigLoader
 CONFIG_LOADER_CLASS = OmegaConfigLoader
 # Keyword arguments to pass to the `CONFIG_LOADER_CLASS` constructor.
 CONFIG_LOADER_ARGS = {
-    "config_patterns": {
-        "snowflake": ["snowflake*"],
-    },
+    "config_patterns": {"snowflake": ["snowflake*"],},
 }
 
 # Class that manages the Data Catalog.